MM-15371: partial implementation to config get/set
This leverages k8s.io/client-go/tools/remotecommand to effectively allow
shell commands on a Mattermost container, and using same to get and set
the config.

It currently relies on the newly introduced --json support to return the
configuration back to the client (and thus a recent enough Mattermost
version), but it can set the configuration on any version.

The approach currently taken is to invoke `mattermost config set ...` on
the first container of the first pod running Mattermost. Ideally, we'd
start our own pod/container to do this, and ideally we wouldn't have to
jump through the hoops that is interacting with the CLI, but could just
hit the API directly. The latter would requires authentication, even for
localhost access.

These changes are missing some unit tests, but I wanted to get this into
review earlier.